backend-django/requirements.txt
# Django et Django REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Base de données PostgreSQL
psycopg2-binary>=2.9.9

# Authentification JWT
djangorestframework-simplejwt>=5.3.0

# Variables d'environnement
python-dotenv>=1.0.0

# Validation et utilitaires
django-filter>=23.5
drf-spectacular>=0.27.0

# Génération de mots de passe sécurisés
secrets
